Always include the Content-Length header
If the client sends a Connection: Close header, Play leaves out the
Content-Length, since it is implicit anyway and computing it may take
some effort in some cases. However, Killing Floor 2 (or more likely
Unreal Engine 3 overall) seems to expect a Content-Length (otherwise the
image does not show up in the server welcome message) *and* it sends a
Connection: Close. So, we always give it a computed Content-Length.